refactor: moves estimateGas and getStorageAt to ContractService (#3662)
